Add RecordError SpanOption (#23)
* Add RecordError SpanOption

RecordError provides an way for consumers to specify if certain errors
should not be recorded on the spans. This can be useful when certain
errors are expected (e.g. a MySQL Duplicate Key, which can be handled
and fall back to an update). Recording errors for these can potentially
bloat the reporting, hiding real errors.

* Add RecordError feature to CHANGELOG

* Add RecordError to README

* Update CHANGELOG.md

Co-authored-by: Sam Xie <xsambundy@gmail.com>

Co-authored-by: Sam Xie <xsambundy@gmail.com>