avoid returning ErrSkip
we observed sql.ErrSkip being returned by stmt.Exec when using
a statement prepared by pq's CopyIn, which returns a statement which
does not implement StmtExecContext.

but the database/sql/driver docs state that "ErrSkip is only supported
where explicitly documented" [1] and StmtExecContext does not explicitly
support it [2].

I looked around a bit to see what other libraries do (dd-trace-go and
nhatthm/otelsql, although I paid the most mind to the former), and it
seems that they choose to fall back to to the contextless equivalents.

This fallback behavior is also exhibited in the standard library [3]
although it's not trivially accessible from a driver and thus I've
reimplemented it herein.

[1] https://pkg.go.dev/database/sql/driver@go1.20.1#pkg-variables
[2] https://pkg.go.dev/database/sql/driver@go1.20.1#StmtExecContext
[3] https://github.com/golang/go/blob/bb8f9a6ae66d742cb67b4ad444179905a537de00/src/database/sql/ctxutil.go#L67