add linter wsl_v5 (#551)
wsl (whitespace linter) is a linter that wants you to use empty lines to
separate grouping of different types to increase readability. There are
also a few places where it encourages you to remove whitespaces which is
at the start and the end of blocks or between assigning and error
checking.